Assert continuity when reassmbling blocks, transactions, and events from StreamEvents.
When serialising BlockExecutions into StreamEvents it is possible that
we do not detect dropped SteamEvents when deserialising (consuming) the SteamEvents later.

This PR makes the StreamEvent consumers, namely: ConsumeBlockExecutions, BlockAccumulator, and TxStack throw errors if there is an incorrect number of events streamed within each container (BlockExecution or TxExecution). To do this a checksum NumTxs and NumEvents is added to BeginBlock and BeginTx respectively.

This allows Vent to crash if it sees an invalid stream (for example due
to messages being dropped due to load on Burrow's event emitter) or a
transport issue. Vent can then restart from its previous good offset.

Annoyance: I had to add in Predecessor to the BlockExecution and
BeginBlock messages in order to track the previous _expected_ block
since we do not store empty blocks in state to avoid changing the AppHash (which itself is just a
workaround for an outstanding issue in tendermint
https://github.com/tendermint/tendermint/issues/1909)

This PR was in response to suspected event drops in Vent in production.

Signed-off-by: Silas Davis <silas@monax.io>